+++ db/migrate/20180803001103_standup_list_init.go
@@ -2,7 +2,7 @@ package migrate
 
 import (
 	"github.com/andrysds/panera/db"
-	"github.com/andrysds/panera/entity/standup"
+	"github.com/andrysds/panera/entity"
 )
 
 func StandupListInit() string {
@@ -22,12 +22,12 @@ func StandupListInit() string {
 		"Luthfi:luthfift:0",
 	}
 
-	if _, err := db.Redis.Del(standup.ListKey).Result(); err != nil {
+	if _, err := db.Redis.Del(entity.StandupListKey).Result(); err != nil {
 		return err.Error()
 	}
 
 	for _, d := range initData {
-		if _, err := db.Redis.RPush(standup.ListKey, d).Result(); err != nil {
+		if _, err := db.Redis.RPush(entity.StandupListKey, d).Result(); err != nil {
 			return err.Error()
 		}
 	}
