+++ db/migrate/20180804031507_standup_init.go
@@ -2,15 +2,15 @@ package migrate
 
 import (
 	"github.com/andrysds/panera/db"
-	"github.com/andrysds/panera/entity/standup"
+	"github.com/andrysds/panera/entity"
 )
 
 func StandupInit() string {
-	if _, err := db.Redis.Del(standup.Key).Result(); err != nil {
+	if _, err := db.Redis.Del(entity.StandupKey).Result(); err != nil {
 		return err.Error()
 	}
 
-	if _, err := db.Redis.Set(standup.Key, 6, 0).Result(); err != nil {
+	if _, err := db.Redis.Set(entity.StandupKey, 6, 0).Result(); err != nil {
 		return err.Error()
 	}
