--- entity/standup/standup_list.go
@@ -1,5 +1,0 @@
-package standup
-
-const ListKey = "panera:standup:list"
-
-type StandupList []*Standup
