+++ handler/standup.go
@@ -5,12 +5,11 @@ import (
 
 	"github.com/andrysds/clarity"
 	"github.com/andrysds/panera/entity"
-	"github.com/andrysds/panera/entity/standup"
 	"gopkg.in/telegram-bot-api.v4"
 )
 
 func HandleStandup(chatID int64) *tgbotapi.MessageConfig {
-	standup, err := standup.Current()
+	standup, err := entity.CurrentStandup()
 	clarity.PrintIfError(err, "error on get standup")
 
 	messageTemplate := "Yuk stand up! Yang dapat giliran untuk memimpin stand up hari ini adalah _%s_ (@%s)"
@@ -21,7 +20,7 @@ func HandleStandup(chatID int64) *tgbotapi.MessageConfig {
 }
 
 func HandleStandupSkip(chatID int64) *tgbotapi.MessageConfig {
-	standup, current, err := standup.Next(false)
+	standup, current, err := entity.NextStandup(false)
 	clarity.PrintIfError(err, "error on skipping standup")
 
 	if err == nil {
