+++ main.go
@@ -11,14 +11,9 @@ import (
 
 func main() {
 	log.Println("Panera starting...")
-
 	config.Init()
 	db.InitRedis()
 	app.Init()
 	cron.Init()
-
-	log.Println("* Listening on tcp://0.0.0.0:" + config.Port)
-	log.Println("Use Ctrl-C to stop")
-
 	app.Panera.Run()
 }
